+++ Models/Product.cs
@@ -18,8 +18,8 @@ namespace StarDMS.Models
         public string Details { get; set; }
         [Required]
         public Guid FirmId { get; set; }
-        public decimal PurchasePrice { get; set; }
-        public decimal SalePrice { get; set; }
+        public int PurchasePrice { get; set; }
+        public int SalePrice { get; set; }
         [Range(0, 100)]
         public int Discount { get; set; }
         public Firm Firm { get; set; }
